# backend/tests/contract/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Medication": {
        "properties": {
          "dosage": {
            "default": "",
            "maxLength": 200,
            "title": "Dosage",
            "type": "string"
          },
          "duration": {
            "default": "",
            "maxLength": 200,
            "title": "Duration",
            "type": "string"
          },
          "form": {
            "default": "",
            "maxLength": 200,
            "title": "Form",
            "type": "string"
          },
          "frequency": {
            "default": "",
            "maxLength": 200,
            "title": "Frequency",
            "type": "string"
          },
          "name": {
            "default": "",
            "maxLength": 200,
            "title": "Name",
            "type": "string"
          }
        },
        "title": "Medication",
        "type": "object"
      },
...
      "PrescriptionCreate": {
        "properties": {
          "additional_notes": {
            "default": "",
            "maxLength": 5000,
            "title": "Additional Notes",
            "type": "string"
          },
          "child_age": {
            "default": "",
            "maxLength": 50,
            "title": "Child Age",
            "type": "string"
          },
          "child_id": {
            "format": "uuid",
            "title": "Child Id",
            "type": "string"
          },
          "child_weight": {
            "default": "",
            "maxLength": 50,
            "title": "Child Weight",
            "type": "string"
          },
          "complaint": {
            "default": "",
            "maxLength": 5000,
            "title": "Complaint",
            "type": "string"
          },
          "date_of_visit": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Visit"
          },
          "diagnosis": {
            "default": "",
            "maxLength": 5000,
            "title": "Diagnosis",
            "type": "string"
          },
          "doctor_name": {
            "default": "",
            "maxLength": 200,
            "title": "Doctor Name",
            "type": "string"
          },
          "medications": {
            "items": {
              "$ref": "#/components/schemas/Medication"
            },
            "title": "Medications",
            "type": "array"
          },
          "source_job_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Job Id"
          },
          "source_text": {
            "default": "",
            "maxLength": 20000,
            "title": "Source Text",
            "type": "string"
          }
        },
        "required": [
          "child_id"
        ],
        "title": "PrescriptionCreate",
        "type": "object"
      },
      "PrescriptionOut": {
        "properties": {
          "additional_notes": {
            "default": "",
            "maxLength": 5000,
            "title": "Additional Notes",
            "type": "string"
          },
          "child_age": {
            "default": "",
            "maxLength": 50,
            "title": "Child Age",
            "type": "string"
          },
          "child_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Child Id"
          },
          "child_weight": {
            "default": "",
            "maxLength": 50,
            "title": "Child Weight",
            "type": "string"
          },
          "complaint": {
            "default": "",
            "maxLength": 5000,
            "title": "Complaint",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "date_of_visit": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Visit"
          },
          "diagnosis": {
            "default": "",
            "maxLength": 5000,
            "title": "Diagnosis",
            "type": "string"
          },
          "doctor_name": {
            "default": "",
            "maxLength": 200,
            "title": "Doctor Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "medications": {
            "items": {
              "$ref": "#/components/schemas/Medication"
            },
            "title": "Medications",
            "type": "array"
          },
          "source_text": {
            "default": "",
            "maxLength": 20000,
            "title": "Source Text",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "PrescriptionOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Multi-user prescription knowledge base with OCR and chat",
    "title": "Medical Prescription Assistant",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/prescriptions": {
      "get": {
        "operationId": "list_prescriptions_api_prescriptions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "child_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Child Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PrescriptionOut"
                  },
                  "title": "Response List Prescriptions Api Prescriptions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Prescriptions",
        "tags": [
          "prescriptions"
        ]
      },
      "post": {
        "operationId": "create_prescription_api_prescriptions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrescriptionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Prescription",
        "tags": [
          "prescriptions"
        ]
      }
    },
    "/api/prescriptions/{prescription_id}": {
      "delete": {
        "operationId": "delete_prescription_api_prescriptions__prescription_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "prescription_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Prescription Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Prescription",
        "tags": [
          "prescriptions"
        ]
      },
      "get": {
        "operationId": "get_prescription_api_prescriptions__prescription_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "prescription_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Prescription Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Prescription",
        "tags": [
          "prescriptions"
        ]
      },
      "put": {
        "operationId": "update_prescription_api_prescriptions__prescription_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "prescription_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Prescription Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrescriptionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Prescription",
        "tags": [
          "prescriptions"
        ]
      }
    },
    "/api/prescriptions/{prescription_id}/photos": {
      "get": {
        "operationId": "get_prescription_photos_api_prescriptions__prescription_id__photos_get",
        "parameters": [
          {
            "in": "path",
            "name": "prescription_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Prescription Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Get Prescription Photos Api Prescriptions  Prescription Id  Photos Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Prescription Photos",
        "tags": [
          "prescriptions"
        ]
      }
    },
...
  }
}